UPDATE: revert/remove playwrights tests UPDATE: Refactor component files/folders
