+++ src/components/organize/people/index.ts
@@ -1,0 +1,5 @@
+export { default as SuggestedViews } from './SuggestedViews';
+
+export { default as CreateViewActionButton } from './CreateViewActionButton';
+
+export { default as ViewsListTable } from './ViewsListTable';
