+++ src/test-utils/mocks/mockView.ts
@@ -1,0 +1,26 @@
+import { mockObject } from 'test-utils/mocks';
+import { ZetkinView } from 'types/zetkin';
+
+const view: ZetkinView = {
+    created: '2021-12-16T14:00:00+00:00',
+    description: 'A random description of this random view',
+    id: 1,
+    organization: {
+        id: 1,
+        title: 'A ',
+    },
+    owner: {
+        id: 1,
+        name: 'M. E. Retsrof',
+    },
+    title: 'A view with a room',
+};
+
+
+const mockView = (overrides?: Partial<ZetkinView>): ZetkinView => {
+    return mockObject(view, overrides);
+};
+
+export default mockView;
+
+export { view };
